refactor: remove login functionality from UserContext and integrate it into LoginForm for better separation of concerns
